requirements.txt
streamlit>=1.32.0
supabase
httpx>=0.27.0,<0.28.0
websockets>=13.0,<16.0
pandas>=2.0.0
python-dotenv
